requirements.txt
babel==2.9.0
python-dateutil==2.6.0
flask-moment==0.11.0
flask-wtf==0.14.3
flask_sqlalchemy==2.4.4
Flask==1.1.2
Jinja2==2.11.3
MarkupSafe==1.1.1
itsdangerous==1.1.0
Werkzeug==1.0.1
Flask-Migrate==3.1.0
alembic==1.7.7
click==8.1.2
Flask-Migrate==3.1.0
greenlet==1.1.2
importlib-metadata==4.11.3
Mako==1.2.0
psycopg2-binary==2.9.3
six==1.16.0
SQLAlchemy==1.4.35
WTForms==3.0.1




